IDEX-3191 Rename CodenvyBootstrap into CheBootstrap and add CodenvyBootstrap as deprecated
